requirements.txt
requests>=2.31.0
feedparser>=6.0.0
google-generativeai>=0.8.0
yfinance>=0.2.40
